Update README.md
price adjustment

$15 price plan for participants under 18 was changed from $25, this is to support any adjustment for the years coming up for the participants. the $25 fee was initially for custom websites but I will adjust that later. 

The $25 plan for anyone over 18 was adjusted from $50, this is due to potential participants who may still pursue STEAM/STEM fields.

This is not an age discrimination, the program is designed for 17 years of age and under with parental approval. Anyone over 18 will be charged the adult fee.